django_project/requirements/base.txt
pytz==2019.1  # https://github.com/stub42/pytz
Pillow==6.0.0  # https://github.com/python-pillow/Pillow
redis==3.2.1  # https://github.com/antirez/redis
celery==4.3.0  # pyup: < 5.0  # https://github.com/celery/celery
django-celery-beat==1.1.1
django-celery-results==1.0.1
flower==0.9.3  # https://github.com/mher/flower
requests==2.22.0  # https://pypi.org/project/requests/

# Django
# ------------------------------------------------------------------------------
Django==2.2  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.1.2  # https://github.com/jazzband/django-model-utils
django-crispy-forms==1.7.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.10.0  # https://github.com/niwinz/django-redis
django-ckeditor==5.7.1  # https://github.com/django-ckeditor/django-ckeditor
django-extensions==2.1.7  # https://github.com/django-extensions/django-extensions
django-smtp-ssl==1.0  # https://github.com/bancek/django-smtp-ssl
django-user-agents==0.3.2  # https://github.com/selwin/django-user_agents
kv-settings==2.0.0  # https://github.com/eshandas/django_key_value_settings
# simple_django_logger==0.1.3

# Django REST Framework
# ------------------------------------------------------------------------------
djangorestframework==3.9.4  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client
django-cors-headers==3.0.1  # https://github.com/ottoyiu/django-cors-headers
coreschema==0.0.4  # https://github.com/core-api/python-coreschema
django-rest-swagger==2.2.0  # https://github.com/marcgibbons/django-rest-swagger